refactor: input setting format.
